Nix expression test + module declaration w/o errs
Testing using a Nix expression to generate the syntax file.

Plain Vim scripts are difficult to maintain once they get large, but Nix
expressions tend to be a bit easier to maintain.

Currently, using the expression is messy, scattered, and poorly
implemented. This will definitely need to be fixed in the future.

As a test of the expression, module declarations were added, minus
support for errors. This also can be improved a lot.